requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.19
sqlalchemy==2.0.36
pyclamd==0.4.0
msoffcrypto-tool==5.4.2
aiofiles==24.1.0
